upload Product and Slide"

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -8,12 +8,16 @@ use App\Models\Category;
 
 class CategoryController extends Controller
 {
+    public function __construct()
+    {
+        date_default_timezone_set("Asia/Ho_Chi_Minh");
+    }
     public function index()
     {
-        $category = Category::all();
+        $categories = Category::orderByDesc('id')->Search()->paginate(5);
         return view('admin.category.list',[
             'title'=>"Danh Sách Danh Mục Mới",
-            'category'=>$category,
+            'categories'=>$categories,
         ]);
     }
 
@@ -21,14 +25,12 @@ class CategoryController extends Controller
     {    
         $this->validate($request,
         [
-            'name'=> 'required|unique:Category,name',
-            
+            'name'=> 'required|unique:Category,name',            
         ],
         [
             'name.required' => 'Vui lòng nhập Tên Danh Mục',
             'name.unique' => 'Tên danh mục đã tồn tại',
         ]);
-        date_default_timezone_set("Asia/Ho_Chi_Minh");
         try{
             Category::create([
                 'name' => (string) $request->input('name'),
@@ -52,8 +54,19 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         try{
-            Category::where('id',$id)->delete();
-            session()->flash('success', 'Xóa Danh Mục Thành Công');
+            //Lấy category có id = $id
+            $category = Category::find($id);
+            //đếm product có id_category = $id
+            $count = $category->product->count();
+            ///session()->flash('success', $count);
+            if($count != 0)
+            {
+                session()->flash('error', 'Danh Mục đang được sử dụng. Xóa không thành công');                
+            }
+            else{
+                Category::where('id',$id)->delete();
+                session()->flash('success', 'Xóa Danh Mục Thành Công');
+            }      
         }catch(\Exception $err){
             session()->flash('error', $err->getMessage());
         }
@@ -62,13 +75,18 @@ class CategoryController extends Controller
     
     public function show($id)
     {
+        //xem chi tiết
+    }
+
+    public function edit($id)
+    {
         $category = Category::find($id);
         return view('admin.category.edit',[
             'title'=>'Chỉnh Sửa Danh Mục: '.$category->name,
             'category'=>$category,
         ]);
     }
-
+    
     public function update(Request $request, $id)
     {
         $category = Category::find($id);
@@ -79,23 +97,12 @@ class CategoryController extends Controller
         [
             'name.required' => 'Vui lòng nhập tên danh mục',
         ]);
-        date_default_timezone_set("Asia/Ho_Chi_Minh");
         try{
-            if($request->input('name') != $category->name 
-            || $request->input('description') != $category->description
-            || $request->input('active') != $category->active)
-            {
-                $category->name = (string) $request->input('name');
-                $category->description = (string) $request->input('description');
-                $category->active = (int) $request->input('active');
-                $category->save();
-                session()->flash('success', 'Cập nhật Danh Mục Thành Công');
-            }
-            else
-            {
-                session()->flash('error', 'Cập nhật Danh Mục Không Thành Công');
-                return redirect('admin/category/edit/'.$id);
-            }           
+            $category->name = (string) $request->input('name');
+            $category->description = (string) $request->input('description');
+            $category->active = (int) $request->input('active');
+            $category->save();
+            session()->flash('success', 'Cập nhật Danh Mục Thành Công');        
         }catch(\Exception $err){
             session()->flash('error', $err->getMessage());
         }
